parseLiteral() function and supporting friends
This adds the "record" struct tag which can be used in record entries.
It is required for all entries of a dnsconfig.Record struct, if you want
to make one optional, add `,optional` to the end of the struct tag's
value.

It recursively converts from LiteralValue, type asserting as it goes by
way of reflect.Value.Set(). It probably has bugs, and still needs
testing.

Signed-off-by: Erik Hollensbe <git@hollensbe.org>